# MDEV-33408 Alter HNSW graph storage and fix memory leak