… sql/vector_mhnsw.cc