@cvicentiu 2024-04-15
Hi @HugoWenTD!

Thank you for the contribution.

I will address your work in order of appearance:

1. The benchmark tool is imensely helpful. I will rebase it at the beggining of bb-11.4-vec-vicentiu for ease of use for all of us. We will have to clean it up a bit, and perhaps it should not reside in the MariaDB Server repo directly, but that's a problem for another time, not relevant during this development effort.

In any case, I have used it to double check your findings and I can reproduce everything (except the results for this particular PR, but more on this later)

2. The "bugfix" you suggested on only deleting one-directional connections: (PR: #3156)
I've re-read the paper a few times and I strongly believe that the graph is only meant to be considered an undirected graph, hence all connections are bi-directional. If you can point out a paragraph to me (or any other source) where the original HNSW allows one-directional links, I am more than happy to reconsider. 

My understanding (at this point in time) is that by "removing only a subset of links", as you've done in #3156 causes a soft increase in MAX_NEIGHBORS_PER_LAYER. The same effect can be achieved by just bumping the MAX_NEIGHBORS_PER_LAYER variable.

So I am inclined to reject that bugfix and instead focus on tuning the graph construction parameters.

3. Here is a set of results I got using the following constants, from the commit 2f6b2b3b7d582b9a483a0ab98448a9ad745d3e51:
```
  const uint EF_CONSTRUCTION = 32; // max candidate list size to connect to.
  const uint MAX_INSERT_NEIGHBOUR_CONNECTIONS = 100;
  const uint MAX_NEIGHBORS_PER_LAYER = 100;
  const double NORMALIZATION_FACTOR = 1.2;
```
without any of your changes in PR #3156
```
  0:    MariaDB(m=16, ef_construction=200, ef_search=10)        0.906     1035.211
```

So we can achieve 90% recall by having more neighbors per layer and bumping eF_construction. Of course, the QPS drops, as the graph is bigger, but as you've noticed, the code was initially written for correctness, not for speed. Once we've proven corectness, we can look into speed improvements.

4. I tried your code from this PR (437e21444ceec6992cd35f4449632a5dec07a9bf). I get the following results, which I believe are due to a bug somewhere:
```
  0:    MariaDB(m=16, ef_construction=200, ef_search=10)        0.005     5095.530`
```

Can you double check your findings and see if you've submitted the correct code to this PR?

In general I agree with your approach of using an adjancency list `(layer, src, neighbors[])` instead of tuples `(layer, src, dst)`.

This approach has some benefits for sure:
1. The number of rows in the table is drastically reduced. (reducing index size)
2. There is less data duplication in the table. (again reducing index size)
3. We are more memory friendly and can thus run searches faster. (I believe we should optimize for SELECT, not for INSERT/UPDATE/DELETE, but of course we can not completely slow down INSERTS)
4. Because of less data to be written to disk, INSERTS should be faster (as your benchmark shows)
5. The adjacency list approach removes a lot of `ha_index_next` calls because it loads all neighbours in memory on one lookup. That's great!
The downsides:
1. Updating a connection will cause a much larger write in the storage engine. We need to benchmark with longer neighbour lists (100 neighbors max for instance).
2. Delete may be trickier to implement. I had started on working on delete with my version of code (layer, src, dst). But I think that your solution has more promise. The way I'm implementing delete is by using a "deleted" bit for a node. If the node is deleted, we skip it during all iterations / searches. I can probably adapt the solution to work with your code. I'll try this right now and share the result as soon as it's working.

5. Your benchmark shows that we need to be able to parameterize the index construction at runtime so we can expriment without recompiling. Would you be interested in tackling that bit? You can create some system variables for now and we'll use those. A quick and dirty solution, just to map with your already written benchmark code. I noticed you have some commented out SET commands that you wanted to use:

```
ann_benchmarks/algorithms/mariadb/module.py
288:        #self._cur.execute("SET hnsw.ef_search = %d" % ef_search)
```

That's all the thoughts I have for now, let me know when you've double checked the benchmark results for your code and in case I messed up when running it, I'll fix it.