@HugoWenTD 2024-04-17
Hi Vicentiu, Thank you for reviewing the pull requests and provide feedback on the benchmarking tool. I really appreciate your input.

I'm traveling for meetings this week but I will try to address on the items you mentioned later this week or next week. You made an important point that the mMax (max neighbours) affects the update speed when using node->neighbours[]. I believe I did run tests with some bigger mMax values. but for the comparison with previous implementation I only recorded the results of the default parameters. I will gather more data on the impact of mMax and get back to you.

---

Regarding your comment:
>  I've re-read the paper a few times and I strongly believe that the graph is only meant to be considered an undirected graph, hence all connections are bi-directional. If you can point out a paragraph to me (or any other source) where the original HNSW allows one-directional links, I am more than happy to reconsider.

Here are a few points supporting the use of one-directional links:

1. The algorithm 1 in the paper describes the shrink steps:  
![image](https://github.com/MariaDB/server/assets/46255328/48b1ccbf-343e-44da-a560-3b93d580ceb1)

At line 16, `set neighbourhood(e) at layer lc to eNewConn` is to shrink only one-direction connection, and it does not mention deleting the connection of the opposite direction.

2. In addition, PgVector's shrinking step is in [this section](https://github.com/pgvector/pgvector/blob/31e41b3ba96a0b8e94062e4f78ea33dee483fd50/src/hnswutils.c#L1073-L1154).  
It also shrinks only one direction.

3. In the worst case, there could be a scenario where vectors gets isolated:
   - Assume mMax-level0 is 4
   - When (V5) vector 5 is inserted, V4 got isolated.
   - If V4 happens to be entry point of higher level. That means none of V1/2/3/5 will be ever traversed in future insert or search.
   - While increasing mMax could mitigate this, it still reduces the navigability of the graph, in the benchmark tests, I did see lower recall caused by fewer vectors were returned than the specified limit.
 
![image](https://github.com/MariaDB/server/assets/46255328/5f7b950d-6726-468a-9dd4-36defe0010c3)